Crate units

Common units used in eviov.

Macros

add_newtype
Defines the operator on the two wrapper types.

add_raw
Defines the operator on a wrapper type and the raw type.

div_newtype
Defines the operator on the two wrapper types.

div_raw
Defines the operator on a wrapper type and the raw type.

mul_newtype
Defines the operator on the two wrapper types.

mul_raw
Defines the operator on a wrapper type and the raw type.

op_newtype
Defines struct Lhs(T) [op] struct Rhs(T).

op_raw
Defines struct Lhs(T) [op] T.

rem_newtype
Defines the operator on the two wrapper types.

rem_raw
Defines the operator on a wrapper type and the raw type.

sub_newtype
Defines the operator on the two wrapper types.

sub_raw
Defines the operator on a wrapper type and the raw type.

Structs

Bearing
An absolute direction bearing, in radians.

Displace
Represents the signed displacement between two positions.

Drag
The drag effect of a fluid.

Ecef
Indicates that the wrapped type is Earth-centered Earth-fixed (ECEF).

Elasticity
The hardness of a solid.

Force
A force vector type.

GameDuration
A non-negative difference between two GameInstants.

GameInstant
A monotonic 50Hz clock in the game system.

Mass
The mass of an object.

Momentum
A momentum vector type.

Position
A Cartesian position, typically using the position of the parent star as the origin.

Rate
The rate value of the wrapped unit T.

Theta
An amount of rotation.

Torque
Represents torque in the counterclockwise direction.

Traits

Direction
A value that represents a direction.

Eci
Structs implementing this trait may represent either ECI or ECEF attributes.

LengthExt
Extension trait for length-specific methods.

Type Definitions

Accel
An acceleration in velocity per GameDuration tick.

Length
A measure of length in a star system.

Omega
An angular speed in radians per GameDuration tick.

Velocity
A velocity in displacement per GameDuration tick.