Type Definition units::Length

type Length = f64;

A measure of length in a star system.

The unit of length is relative to the star system. When using length values from two different star systems, scaling must first be performed.

TODO change to newtype